fix(WPN2d): clippy at -D warnings, three findings
* `collapsible_match` in the inventory panel's bench arm. The two lookups are
  `let`s and not nested `if`s now, deliberately: a match GUARD would put the
  condition on its own line, and that is where
  `every_key_the_inventory_panel_reads_has_a_route` scrapes the reducer's key
  literals from -- so the arm clippy asks for would have made the arm that
  found the unrouted keys blind again.
* `unnecessary_min_or_max` on the magazine top-up: `u32::max(0)` is a no-op.
* `useless_conversion` on the pellet refusal, and `too_many_arguments` on
  `resolve_swing`, which is eight because the LOS probe needs the bridge and the
  counter needs the report -- allowed by name beside `resolve_shot`, which
  carries the same allow for the same reason.

Clippy `--workspace --all-targets` at `-D warnings`: **0 errors, 0 warnings**.
The lib tests of every crate this commit touches re-run green.

+++ crates/inf-physics/src/d3/gameplay.rs
@@ -1937,7 +1937,7 @@ fn step_weapons(
             // one, which `resolve_shot` keeps.
             let bill = report.rounds.shot_rays as usize + 1;
             if bill > inf_ecs::ballistics::MAX_SHOT_RAYS_PER_STEP {
-                report.rounds.pellets_refused += u32::from(pellets - p);
+                report.rounds.pellets_refused += pellets - p;
                 break;
             }
             // Consecutive counter indices, strided by the pellet bound so two
@@ -3324,6 +3324,7 @@ fn strike_point(world: &EcsWorld, guid: Uuid) -> Option<DVec3> {
 /// `O(characters)`, over the same walk [`gunners`] already makes — and only on
 /// the steps a swing actually leaves, which at [`weapon::FIST_RPM`] is at most
 /// one and a half a second.
+#[allow(clippy::too_many_arguments)]
 fn resolve_swing(
     world: &EcsWorld,
     bridge: &mut PhysicsBridge3D,
